+++ .eslintrc-node.js
@@ -1,0 +1,9 @@
+module.exports = {
+	rules: {
+		'node/handle-callback-err': 2,
+		'node/no-callback-literal': 2,
+		'node/no-new-require': 2,
+		'node/no-path-concat': 2,
+		'node/no-process-exit': 2,
+	},
+};