+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+## 7.16-r0 / 2020-12-20
+* Update rules for ESLint 7.16.
+* Add module for **eslint-plugin-node**.
+
 ## 6.8-r0 / 2020-12-20
 * Add baseline rules for the last ESLint version to support Node.js 8 LTS.